added pins for min endstops
now you can use max or min end stops for all axis. You only have to
change the homing direction corresponding to your build. In future I
will change the homing direction of X and Y to the min end stop since
this seems to be the official position now.